% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level2_validation.R
\name{adjust_fact_plot}
\alias{adjust_fact_plot}
\title{Plot adjustment factor}
\usage{
adjust_fact_plot(
  nsa,
  sa,
  title = NULL,
  easter_lag = 6,
  julian_easter = FALSE,
  default_type = "X13",
  default_spec_nsa = "RSA2c"
)
}
\arguments{
\item{nsa}{ts object (usually unadjusted time series)}

\item{sa}{ts object (usually seasonally adjusted time series)}

\item{title}{optional title}

\item{easter_lag}{integer defining the number of days before Easter Sunday to create the Easter window}

\item{julian_easter}{logical, should EAster be a Julian Easter, default FALSE}

\item{default_type}{values should be "X13" or "TS" to define whether X13 or TRAMO-SEATS is used to test for decomposition mode}

\item{default_spec_nsa}{name of a default JDemetra+ specification to use for determining decomposition mode (default is  "RSA2c")}
}
\value{
A plot of derived adjustment factors
}
\description{
Plot adjustment factor
}
\examples{
data(data_to_check, package = "SAvalidation")
adjust_fact_plot(data_to_check$nsa,data_to_check$sa)
}
