% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level2_validation.R
\name{annual_totals_plot}
\alias{annual_totals_plot}
\title{Plot of relative difference of annual totals}
\usage{
annual_totals_plot(nsa, sa, title = NULL)
}
\arguments{
\item{nsa}{ts object (usually unadjusted time series)}

\item{sa}{ts object (usually seasonally adjusted time series)}

\item{title}{optional title}
}
\value{
A plot of the relative difference of annual totals between nsa and sa
}
\description{
Plot of relative difference of annual totals
}
\examples{
load(data_to_check)
annual_totals_plot(data_to_check$nsa,data_to_check$sa)
}
