% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level2_validation.R
\name{cal_effect_plot}
\alias{cal_effect_plot}
\title{Plot of calendar effects}
\usage{
cal_effect_plot(
  nsa,
  sa,
  sa_mod,
  title = NULL,
  default_type = "X13",
  default_spec_nsa = "RSA2c"
)
}
\arguments{
\item{nsa}{ts object (usually unadjusted time series)}

\item{sa}{ts object (usually seasonally adjusted time series)}

\item{sa_mod}{An SA object from RJDemetra}

\item{title}{optional title}

\item{default_type}{values should be "X13" or "TS" to define whether X13 or TRAMO-SEATS is used to test for decomposition mode}

\item{default_spec_nsa}{name of a default JDemetra+ specification to use for determining decomposition mode (default is  "RSA2c")}
}
\value{
A plot of calendar effects and derived adjusted effects
}
\description{
Plot of calendar effects
}
\examples{
test_sa_mod <- RJDemetra::x13(data_to_check$sa,
spec="RSA2c",
userdefined = RJDemetra::user_defined_variables("X13-ARIMA"))
cal_effect_plot(data_to_check$nsa,data_to_check$sa,test_sa_mod)
}
