% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level1_validation.R
\name{check_for_calendar_vars}
\alias{check_for_calendar_vars}
\title{Extract test for calendar effects}
\usage{
check_for_calendar_vars(sa_mod)
}
\arguments{
\item{sa_mod}{An SA object from RJDemetra}
}
\value{
A logical indicating whether calendar regressors found in the SA model object
}
\description{
Extract test for calendar effects
}
\examples{
data(data_to_check)
nsa <- data_to_check$nsa
jd_mod <- RJDemetra::x13(nsa)
check_for_calendar_vars(jd_mod)
}
