% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level1_validation.R
\name{check_identical}
\alias{check_identical}
\title{Check if two series are identical}
\usage{
check_identical(nsa, sa)
}
\arguments{
\item{nsa}{A ts object}

\item{sa}{A ts object}
}
\value{
Logical TRUE if all(nsa==sa)
}
\description{
Check if two series are identical
}
\examples{
data(data_to_check)
check_identical(data_to_check$nsa,data_to_check$sa)
}
