% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level1_validation.R
\name{check_negatives}
\alias{check_negatives}
\title{Check for negatives}
\usage{
check_negatives(series_to_check)
}
\arguments{
\item{series_to_check}{A numeric vector (or matrix) to check for any negative values}
}
\value{
Logical TRUE if any negative values exist
}
\description{
Check for negatives
}
\examples{
data(data_to_check)
check_negatives(data_to_check$nsa)
}
