% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level2_validation.R
\name{check_nsa_sa_ts}
\alias{check_nsa_sa_ts}
\title{Check nsa and sa are univariate time series of same period}
\usage{
check_nsa_sa_ts(nsa, sa)
}
\arguments{
\item{nsa}{ts object}

\item{sa}{ts object}
}
\value{
stops if problem with nsa and sa time series
}
\description{
Check nsa and sa are univariate time series of same period
}
\examples{
data(data_to_check)
check_nsa_sa_ts(data_to_check$nsa,data_to_check$sa)
}
