% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level1_validation.R
\name{check_over_adjustment}
\alias{check_over_adjustment}
\title{Check for over adjustment}
\usage{
check_over_adjustment(sa_mod, pval = 0.05)
}
\arguments{
\item{sa_mod}{An SA object from RJDemetra generated with userdefined="preprocessing.model.y_lin"}

\item{pval}{numeric value used to define a p-value to use as a threshold for accepting/rejecting null hypothesis (default 0.05)}
}
\value{
Logical evidence of
}
\description{
Tests null hypothesis that acf at lag 4 is greater or equal to zero
}
\examples{
data(data_to_check)
sa <- data_to_check$sa
jd_mod <- RJDemetra::x13(sa)
check_over_adjustment(jd_mod)
}
