% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level1_validation.R
\docType{data}
\name{data_to_check}
\alias{data_to_check}
\title{Example data for validation}
\format{
list object with name, nsa and sa time series
}
\description{
Example data for validation
}
\keyword{data}
