% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level1_validation.R
\name{get_combined_seasonality_test}
\alias{get_combined_seasonality_test}
\title{Extract combined test for seasonality on linearised series}
\usage{
get_combined_seasonality_test(sa_mod)
}
\arguments{
\item{sa_mod}{An SA object from RJDemetra}
}
\value{
A character vector stating the final result of the combined test for seasonality
}
\description{
Extract combined test for seasonality on linearised series
}
\examples{
data(data_to_check)
nsa <- data_to_check$nsa
jd_mod <- RJDemetra::jx13(nsa)
get_combined_seasonality_test(jd_mod)
}
