% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level1_validation.R
\name{level1_validation}
\alias{level1_validation}
\title{Level 1 validation check}
\usage{
level1_validation(
  nsa,
  sa,
  default_type = "X13",
  default_spec_nsa = "RSA1",
  default_spec_sa = "RSA2c"
)
}
\arguments{
\item{nsa}{A ts object}

\item{sa}{A ts object}

\item{default_type}{Character must be either "X13" (default) or "TS" determining whether X13 or TRAMO-SEATS is used for testing}

\item{default_spec_nsa}{Character defining the JDemetra+ specification for tests on the nsa series (default="RSA1")}

\item{default_spec_sa}{Character defining the JDemetra+ specification for tests on the sa series (default="RSA2c")}
}
\value{
Message about the level 1 validation (pass, pass with warnings or fail)
}
\description{
Level 1 validation check
}
\examples{
data(data_to_check)
level1_validation(data_to_check$nsa,data_to_check$sa)
}
