% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level2_validation.R
\name{level2_validation}
\alias{level2_validation}
\title{Title}
\usage{
level2_validation(
  nsa,
  sa,
  series_name,
  output_directory = NULL,
  dashboard_template = "skeleton.qmd",
  start_date = "1999-01-01",
  default_type = "X13",
  default_spec_nsa = "RSA2c",
  default_spec_sa = "RSA2c",
  java_home = Sys.getenv("JAVA_HOME")
)
}
\arguments{
\item{nsa}{ts object (usually unadjusted time series)}

\item{sa}{ts object (usually seasonally adjusted time series)}

\item{series_name}{a name for the time series to be analysed}

\item{output_directory}{optional output directory for dashboard (default uses getwd())}

\item{dashboard_template}{name of dashboard template to use}

\item{start_date}{Character defining start date in format "YYYY-MM-DD"}

\item{default_type}{values should be "X13" or "TS" to define whether X13 or TRAMO-SEATS is used to test for decomposition mode}

\item{default_spec_nsa}{name of a default JDemetra+ specification to use for tests on NSA series (default is  "RSA2c")}

\item{default_spec_sa}{name of a default JDemetra+ specification to use for tests on SA series (default is  "RSA2c")}

\item{java_home}{JAVA_HOME environment variable}
}
\value{
creates an html dashboard with series name in given output directory
}
\description{
Title
}
\examples{
\donttest{
data(data_to_check, package = "SAvalidation")
level2_validation(data_to_check$nsa,data_to_check$sa,data_to_check$name)
}
}
