% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level3_validation.R
\name{level3_comparisons_plot}
\alias{level3_comparisons_plot}
\title{Level 3  comparison plot}
\usage{
level3_comparisons_plot(
  nsa,
  sa,
  series_name,
  default_spec_nsa = "RSA1",
  test_level = 0.01,
  java_home = Sys.getenv("JAVA_HOME")
)
}
\arguments{
\item{nsa}{ts object (usually unadjusted time series)}

\item{sa}{ts object (usually seasonally adjusted time series)}

\item{default_spec_nsa}{name of a default JDemetra+ specification to use for determining decomposition mode (default is  "RSA1")}

\item{test_level}{significance level for tests (default is 0.01)}

\item{java_home}{}

\item{sa_mod}{An SA object from RJDemetra}
}
\value{
a ggplot object showing the published SA series and some approximate confidence intervals
}
\description{
Level 3  comparison plot
}
