% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level3_validation.R
\name{level3_validation}
\alias{level3_validation}
\title{Level 3 validation}
\usage{
level3_validation(
  nsa,
  sa,
  series_name,
  vintages = NULL,
  dataset_name = NULL,
  title = series_name,
  output_directory = NULL,
  dashboard_template = "skeleton3.qmd",
  start_date = "1999-01-01",
  default_type = "X13",
  default_spec_nsa = "RSA2c",
  default_spec_sa = "RSA2c",
  java_home = Sys.getenv("JAVA_HOME")
)
}
\arguments{
\item{nsa}{ts object (usually unadjusted time series)}

\item{sa}{ts object (usually seasonally adjusted time series)}

\item{series_name}{a name for the time series to be analysed}

\item{vintages}{a list with two elements, nsa_vert and sa_vert, each of which is a matrix with columns as dates and rows as vintages}

\item{dataset_name}{a name for the dataset}

\item{title}{title}

\item{output_directory}{optional output directory for dashboard (default uses getwd())}

\item{dashboard_template}{name of dashboard template to use}

\item{start_date}{Character defining start date in format "YYYY-MM-DD"}

\item{default_type}{values should be "X13" or "TS" to define whether X13 or TRAMO-SEATS is used to test for decomposition mode}

\item{default_spec_nsa}{name of a default JDemetra+ specification to use for tests on NSA series (default is  "RSA2c")}

\item{default_spec_sa}{name of a default JDemetra+ specification to use for tests on SA series (default is  "RSA2c")}

\item{java_home}{JAVA_HOME environment variable}
}
\value{
creates an html document with series name in given output directory
}
\description{
Level 3 validation
}
\examples{
\donttest{
data(vintages, package = "SAvalidation")
nsa <- ts(vintages$nsa_vert[,"2024-01-01"], start=1999, frequency = 4)
sa <- ts(vintages$sa_vert[,"2024-01-01"], start=1999, frequency = 4)
level3_validation(nsa,sa,"test",vintages)
}
}
